Fixes #22

+++ OBA/online_bulletin.rpy
@@ -444,8 +444,9 @@ label splashscreen:
     return
 
 label quit:
-    $ fetch_rpy(persistent.current_news["bullet_file"])
-    $ delete_file()
+    if persistent.current_news != None:
+        $ fetch_rpy(persistent.current_news["bullet_file"])
+        $ delete_file()
     $ persistent.news = None
     $ persistent.current_news = None
     $ cleanup_temp_files()
